Use common lisp functions
`caddr` is not a Elisp function, which leads to the error:

"Company: backend company-reftex-labels error "Symbol’s function definition is void: caddr" with args"

Change `caddr` to `cl-addr`. The former is declared in `cl.el` which
is not `require`d in this package. Since `cl-caddr` is a function, we
need to promote `cl-lib` to a run-time dependency.